pyproject.toml
[tool.poetry]
name = "support-ticket-agent"
version = "0.1.0"
description = "Support ticket agent"
authors = ["Chavvi Bhatia <chavvibhatia23@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain-google-genai = "^4.1.3"
langgraph = "^0.6.3"
langchain-core = "^1.2.6"
langgraph-cli = {extras = ["inmem"], version = "^0.4.11"}
python-dotenv = "^1.0.0"
langchain-community = "^0.4.1"
langchain-huggingface = "^1.2.0"
huggingface-hub = "^0.33.4"
[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
